split the desc apis, test=develop
