+++ lite/model_parser/naive_buffer/block_desc.h
@@ -56,6 +56,11 @@ class BlockDesc : public BlockDescAPI {
   T* GetVar(int32_t idx);
 
   template <typename T>
+  T const* GetVar(int32_t idx) const {
+    return GetVar<T>(idx);
+  }
+
+  template <typename T>
   T* AddVar();
 
   size_t OpsSize() const override;
@@ -64,6 +69,11 @@ class BlockDesc : public BlockDescAPI {
 
   template <typename T>
   T* GetOp(int32_t idx);
+
+  template <typename T>
+  T const* GetOp(int32_t idx) const {
+    return GetOp<T>(idx);
+  }
 
   template <typename T>
   T* AddOp();
