+++ lite/model_parser/naive_buffer/program_desc.h
@@ -46,6 +46,11 @@ class ProgramDesc : public ProgramDescAPI {
   T *GetBlock(int32_t idx);
 
   template <typename T>
+  T const *GetBlock(int32_t idx) const {
+    return GetBlock<T>(idx);
+  }
+
+  template <typename T>
   T *AddBlock();
 
   bool HasVersion() const override { return true; }
