+++ lite/model_parser/pb/block_desc.h
@@ -51,6 +51,11 @@ class BlockDesc : public BlockDescAPI {
   T* GetVar(int32_t idx);
 
   template <typename T>
+  T const* GetVar(int32_t idx) const {
+    return GetVar<T>(idx);
+  }
+
+  template <typename T>
   T* AddVar();
 
   size_t OpsSize() const override { return desc_->ops_size(); }
@@ -59,6 +64,11 @@ class BlockDesc : public BlockDescAPI {
 
   template <typename T>
   T* GetOp(int32_t idx);
+
+  template <typename T>
+  T const* GetOp(int32_t idx) const {
+    return GetOp<T>(idx);
+  }
 
   template <typename T>
   T* AddOp();
