+++ lite/model_parser/pb/program_desc.h
@@ -43,6 +43,11 @@ class ProgramDesc : public ProgramDescAPI {
   T *GetBlock(int32_t idx);
 
   template <typename T>
+  T const *GetBlock(int32_t idx) const {
+    return GetBlock<T>(idx);
+  }
+
+  template <typename T>
   T *AddBlock();
 
   bool HasVersion() const override { return desc_->has_version(); }
